feat: controller와 view 구현
