--- src/main/java/card/DeckFactory.java
@@ -1,24 +1,0 @@
-package card;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class DeckFactory {
-    public static List<Card> create() {
-        List<Card> cards = new ArrayList<>();
-        Symbol[] symbols = Symbol.values();
-        for (Symbol symbol : symbols) {
-            createByType(cards, symbol);
-        }
-        Collections.shuffle(cards);
-        return cards;
-    }
-
-    private static void createByType(List<Card> cards, Symbol symbol) {
-        Type[] types = Type.values();
-        for (Type type : types) {
-            cards.add(new Card(symbol, type));
-        }
-    }
-}
