--- src/main/java/card/Type.java
@@ -1,8 +1,0 @@
-package card;
-
-public enum Type {
-    SPADE,
-    DIAMOND,
-    HEART,
-    CLUB
-}
