+++ src/main/java/domains/card/Type.java
@@ -1,0 +1,8 @@
+package domains.card;
+
+public enum Type {
+    SPADE,
+    DIAMOND,
+    HEART,
+    CLUB
+}
