+++ src/main/java/domains/user/InvalidPlayerException.java
@@ -1,0 +1,10 @@
+package domains.user;
+
+public class InvalidPlayerException extends IllegalArgumentException {
+    public static final String INVALID_INPUT = "y 혹은 n으로 입력해주세요.";
+    public static final String NULL_OR_EMPTY = "빈 문자열이나 null은 입력할 수 없습니다.";
+
+    public InvalidPlayerException(String s) {
+        super(s);
+    }
+}
