--- src/main/java/user/Dealer.java
@@ -1,22 +1,0 @@
-package user;
-
-import card.Deck;
-
-public class Dealer extends User {
-    public static final int DEALER_HIT_POINT = 16;
-
-    public Dealer(Deck deck) {
-        this.hands = new Hands(deck);
-    }
-
-    public Dealer(Hands hands) {
-        this.hands = hands;
-    }
-
-    @Override
-    public void hit(Deck deck) {
-        if (this.hands.score() <= DEALER_HIT_POINT) {
-            this.hands.draw(deck);
-        }
-    }
-}
