--- src/main/java/user/InvalidPlayerException.java
@@ -1,10 +1,0 @@
-package user;
-
-public class InvalidPlayerException extends IllegalArgumentException {
-    public static final String INVALID_INPUT = "y 혹은 n으로 입력해주세요.";
-    public static final String NULL_OR_EMPTY = "빈 문자열이나 null은 입력할 수 없습니다.";
-
-    public InvalidPlayerException(String s) {
-        super(s);
-    }
-}
